Integrated version 1.0

+++ android/app/src/main/java/com/yys/cs446/es/castle_model/settler.java
@@ -2,7 +2,7 @@ package com.yys.cs446.es.castle_model;
 
 public class settler extends unit{
 
-	settler(int x, int y, player p) {
+	settler(double x, double y, player p) {
 		super(x, y, p);
 	}
